# swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
    "/auth/register": {
      "post": {
        "operationId": "AuthController_register",
        "summary": "Register User",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/RegisterDto" }
            }
          }
        },
        "responses": {
          "201": { "description": "User registration." },
          "401": { "description": "User already exists." }
        },
        "tags": ["Auth"]
      }
    },
    "/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "summary": "Log user",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/LoginDto" }
            }
          }
        },
        "responses": {
          "200": { "description": "User Login." },
          "401": { "description": "Invalid credentials." }
        },
        "tags": ["Auth"]
      }
    },
    "/books": {
      "get": {
        "operationId": "BooksController_getAllProducts",
        "summary": "Get all products.",
        "parameters": [
          {
            "name": "page",
            "required": true,
            "in": "query",
            "schema": { "type": "number" }
          },
          {
            "name": "take",
            "required": true,
            "in": "query",
            "schema": { "type": "number" }
          }
        ],
        "responses": { "200": { "description": "Getting all products." } },
        "tags": ["Books"],
        "security": [{ "bearer": [] }]
      },
      "post": {
        "operationId": "BooksController_addProduct",
        "summary": "Insert product",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/BookDto" }
            }
          }
        },
        "responses": {
          "201": { "description": "Insert product." },
          "401": { "description": "User not logged in." }
        },
        "tags": ["Books"],
        "security": [{ "bearer": [] }, { "bearer": [] }]
      }
    },
    "/books/{IBSN}": {
      "delete": {
        "operationId": "BooksController_removeProduct",
        "summary": "Delete product",
        "parameters": [
          {
            "name": "IBSN",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": { "description": "Product deleted." },
          "401": { "description": "User not logged in." }
        },
        "tags": ["Books"],
        "security": [{ "bearer": [] }, { "bearer": [] }]
      }
    },
    "/books/price/{IBSN}": {
      "patch": {
        "operationId": "BooksController_changePriceOfBook",
        "summary": "Change price of book",
        "parameters": [
          {
            "name": "IBSN",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/BookPriceDto" }
            }
          }
        },
        "responses": {
          "200": { "description": "Price Changed." },
          "401": { "description": "User not logged in." }
        },
        "tags": ["Books"],
        "security": [{ "bearer": [] }, { "bearer": [] }]
      }
    },
    "/books/review": {
      "post": {
        "operationId": "BooksController_addReviewToBook",
        "summary": "Place review for book",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ReviewDto" }
            }
          }
        },
        "responses": {
          "200": { "description": "Review placed." },
          "401": { "description": "User not logged in." }
        },
        "tags": ["Books"],
        "security": [{ "bearer": [] }, { "bearer": [] }]
      }
    },
    "/books/review/{IBSN}": {
      "delete": {
        "operationId": "BooksController_removeReviewForBook",
        "summary": "Delete review for book",
        "parameters": [
          {
            "name": "IBSN",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/DeleteReviewDto" }
            }
          }
        },
        "responses": {
          "200": { "description": "Review deleted." },
          "401": { "description": "User not logged in." }
        },
        "tags": ["Books"],
        "security": [{ "bearer": [] }, { "bearer": [] }]
      }
    },
    "/basket/add": {
      "post": {
        "operationId": "BasketController_insertProductToBasket",
        "summary": "Insert product in basket",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/BasketBook" }
            }
          }
        },
        "responses": {
          "200": { "description": "Product inserted in basket." },
          "401": { "description": "User not logged in." }
        },
        "tags": ["Basket"],
        "security": [{ "bearer": [] }]
      }
    },
    "/basket": {
      "get": {
        "operationId": "BasketController_addProduct",
        "summary": "Get all baskets",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/BasketDto" }
            }
          }
        },
        "responses": {
          "201": { "description": "show list of baskets" },
          "401": { "description": "User not logged in." }
        },
        "tags": ["Basket"],
        "security": [{ "bearer": [] }]
      }
    },
    "/basket/{IBSN}": {
      "delete": {
        "operationId": "BasketController_removeProduct",
        "summary": "Delete product from basket",
        "parameters": [
          {
            "name": "IBSN",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": { "description": "Product deleted from basket." },
          "401": { "description": "User not logged in." }
        },
        "tags": ["Basket"],
        "security": [{ "bearer": [] }]
      },
      "patch": {
        "operationId": "BasketController_updateCountOfBookInBasket",
        "summary": "Update count of book in user basket",
        "parameters": [
          {
            "name": "IBSN",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/BasketBookDto" }
            }
          }
        },
        "responses": {
          "200": { "description": "Count of book changed" },
          "401": { "description": "User not logged in." }
        },
        "tags": ["Basket"],
        "security": [{ "bearer": [] }]
      }
    },
    "/basket/books": {
      "get": {
        "operationId": "BasketController_showBooksInBasket",
        "summary": "Show all books in user basket",
        "parameters": [],
        "responses": {
          "200": { "description": "show list" },
          "401": { "description": "User not logged in." }
        },
        "tags": ["Basket"],
        "security": [{ "bearer": [] }]
      }
    },
    "/order": {
      "post": {
        "operationId": "OrderController_addBasketToOrder",
        "summary": "Create order for logged user",
        "parameters": [],
        "responses": {
          "200": { "description": "Order created." },
          "401": { "description": "User not logged in." }
        },
        "tags": ["Order"],
        "security": [{ "bearer": [] }]
      },
      "get": {
        "operationId": "OrderController_getAllOrders",
        "summary": "Show list of placed orders",
        "parameters": [],
        "responses": {
          "200": { "description": "Order list shown" },
          "401": { "description": "User not logged in." }
        },
        "tags": ["Order"],
        "security": [{ "bearer": [] }]
      }
    }
  },
  "info": {
    "title": "Project-sql",
    "description": "The Project-sql API description",
    "version": "1.0",
    "contact": {}
  },
  "tags": [{ "name": "Endpoints", "description": "" }],
  "servers": [],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "Token",
        "type": "http"
      }
    },
    "schemas": {
      "RegisterDto": {
        "type": "object",
        "properties": {
          "email": { "type": "string", "description": "email" },
          "password": { "type": "string", "description": "password" },
          "name": { "type": "string", "description": "name" }
        },
        "required": ["email", "password", "name"]
      },
      "LoginDto": {
        "type": "object",
        "properties": {
          "email": { "type": "string", "description": "email" },
          "password": { "type": "string", "description": "password" }
        },
        "required": ["email", "password"]
      },
      "BookDto": {
        "type": "object",
        "properties": {
          "IBSN": { "type": "number", "description": "IBSN" },
          "title": { "type": "string", "description": "Title of book" },
          "price": { "type": "number", "description": "price of book" },
          "author": { "type": "string", "description": "Name of author" },
          "count": { "type": "number", "description": "count of book" }
        },
        "required": ["IBSN", "title", "price", "author", "count"]
      },
      "BookPriceDto": {
        "type": "object",
        "properties": {
          "price": { "type": "number", "description": "price of book" }
        },
        "required": ["price"]
      },
      "ReviewDto": { "type": "object", "properties": {} },
      "DeleteReviewDto": { "type": "object", "properties": {} },
      "BasketBook": { "type": "object", "properties": {} },
      "BasketDto": {
        "type": "object",
        "properties": {
          "basketId": {
            "type": "string",
            "description": "Basket_id",
            "nullable": false
          },
          "userId": {
            "type": "string",
            "description": "User_id",
            "nullable": false
          }
        },
        "required": ["basketId", "userId"]
      },
      "BasketBookDto": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "description": "id" },
          "count": { "type": "string", "description": "Count" }
        },
        "required": ["id", "count"]
      }
    }
  }
}